requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
filelock==3.9.0
firebase==4.0.1
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
platformdirs==3.1.1
psycopg2-binary==2.9.6
pytz==2023.2
requests==2.29.0
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.20.0
whitenoise==6.5.0
